+++ code/src/cljs/sixsq/nuvla/ui/credentials/subs.cljs
@@ -1,6 +1,6 @@
 (ns sixsq.nuvla.ui.credentials.subs
   (:require
-    [re-frame.core :refer [reg-sub]]
+    [re-frame.core :refer [reg-sub subscribe]]
     [sixsq.nuvla.ui.credentials.spec :as spec]))
 
 
@@ -79,3 +79,56 @@
   ::credential-modal-visible?
   (fn [db]
     (::spec/credential-modal-visible? db)))
+
+
+(reg-sub
+  ::credential-check-table
+  (fn [db]
+    (::spec/credential-check-table db)))
+
+(reg-sub
+  ::credential-check
+  (fn [db [_ id]]
+    (get-in db [::spec/credential-check-table id])))
+
+
+(reg-sub
+  ::credential-check-loading?
+  (fn [[_ id]] (subscribe [::credential-check id]))
+  (fn [credential-check]
+    (boolean (:check-in-progress? credential-check))))
+
+
+(reg-sub
+  ::credential-check-last-check
+  (fn [[_ id]] (subscribe [::credential-check id]))
+  (fn [credential-check]
+    (:last-check credential-check)))
+
+
+(reg-sub
+  ::credential-check-error-msg
+  (fn [[_ id]] (subscribe [::credential-check id]))
+  (fn [credential-check]
+    (:error-msg credential-check)))
+
+
+(reg-sub
+  ::credential-check-status
+  (fn [[_ id]] (subscribe [::credential-check id]))
+  (fn [credential-check]
+    (:status credential-check)))
+
+
+(reg-sub
+  ::credential-check-status-invalid?
+  (fn [[_ id]] (subscribe [::credential-check-status id]))
+  (fn [status]
+    (= status "INVALID")))
+
+
+(reg-sub
+  ::credential-check-status-valid?
+  (fn [[_ id]] (subscribe [::credential-check-status id]))
+  (fn [status]
+    (= status "VALID")))